Put getStatus function in Runner class
